--- addons/l10n_id_efaktur/__init__.py
@@ -1,6 +1,0 @@
-# -*- encoding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import models
-from . import controllers
-from . import wizard
