--- addons/l10n_id_efaktur/controllers/__init__.py
@@ -1,1 +1,0 @@
-from . import download_efaktur
