--- addons/l10n_id_efaktur/models/__init__.py
@@ -1,7 +1,0 @@
-# -*- encoding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import efaktur
-from . import account_move
-from . import res_partner
-from . import efaktur_document
