--- addons/l10n_id_efaktur/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_efaktur_user,access.efaktur.user,model_l10n_id_efaktur_efaktur_range,account.group_account_invoice,1,1,1,1
-access_efaktur_document_user,access.efkatur.document.user,model_l10n_id_efaktur_document,account.group_account_invoice,1,1,1,1
