--- addons/l10n_id_efaktur/tests/__init__.py
@@ -1,2 +1,0 @@
-from . import test_l10n_id_efaktur
-from . import test_l10n_id_efaktur_download
