--- addons/l10n_id_efaktur/views/efaktur_views.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <record id="efaktur_tree_view" model="ir.ui.view">
-            <field name="name">l10n_id_efaktur.efaktur.range.list.view</field>
-            <field name="model">l10n_id_efaktur.efaktur.range</field>
-            <field name="arch" type="xml">
-                <list string="Efaktur Number" editable="bottom">
-                    <field name="min"/>
-                    <field name="max"/>
-                    <field name="available" sum="Total Available"/>
-                    <field name="company_id" groups="base.group_multi_company"/>
-                    <field name="company_id" column_invisible="True" groups="!base.group_multi_company"/>
-                </list>
-            </field>
-        </record>
-
-        <record id='efaktur_invoice_action' model='ir.actions.act_window'>
-            <field name="name">e-Faktur Ranges</field>
-            <field name="res_model">l10n_id_efaktur.efaktur.range</field>
-            <field name="view_mode">list</field>
-            <field name="context">{'search_default_upload': True, 'search_default_used': True}</field>
-            <field name="view_id" ref="efaktur_tree_view"/>
-            <field name="help" type="html">
-                <p class="oe_view_nocontent_create">
-                    Put here the ranges of number you were assigned by the government. Those
-                    will be assigned to your Customer Invoices upon confirmation.
-                </p>
-            </field>
-        </record>
-
-
-    </data>
-</odoo>
