--- addons/l10n_id_efaktur/views/res_config_settings_views.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="res_config_settings_view_form" model="ir.ui.view">
-        <field name="name">res.config.settings.view.form.chilean.loc</field>
-        <field name="model">res.config.settings</field>
-        <field name="inherit_id" ref="account.res_config_settings_view_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//block[@id='invoicing_settings']" position="after">
-                <block title="Indonesian Localization" id="l10n_cl_section" invisible="country_code != 'ID'">
-                    <setting id="l10n_id_efaktur">
-                        <div class="content-group">
-                            <button type="action" name="%(l10n_id_efaktur.efaktur_invoice_action)d" string="Configure Your e-Faktur Ranges" class="btn-link" icon="oi-arrow-right"/>
-                        </div>
-                    </setting>
-                </block>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
