--- addons/l10n_id_efaktur/views/res_partner_views.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <record id="res_partner_tax_form_view" model="ir.ui.view">
-            <field name="name">res.partner.tax.form.view</field>
-            <field name="model">res.partner</field>
-            <field name="inherit_id" ref="base.view_partner_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//group/group" position="inside">
-                    <field name="l10n_id_pkp" invisible="country_id and country_code != 'ID'"/>
-                    <field name="l10n_id_kode_transaksi" invisible="not l10n_id_pkp"/>
-                </xpath>
-                <page name="accounting" position="inside">
-                    <group string="Indonesian Taxes"  invisible="not l10n_id_pkp">
-                        <group>
-                            <field name="l10n_id_nik"/>
-                        </group>
-                    </group>
-                </page>
-            </field>
-        </record>
-    </data>
-</odoo>
