--- addons/l10n_id_efaktur/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import account_move_reversal
