+++ addons/l10n_id_efaktur_coretax/__manifest__.py
@@ -13,7 +13,7 @@
         is resulting to 11%.
     """,
     'category': 'Accounting/Localizations/EDI',
-    'depends': ['l10n_id', 'l10n_id_efaktur'],
+    'depends': ['l10n_id'],
     'data': [
         # New Data Import (E-Faktur code related)
         "data/l10n_id_efaktur_coretax.product.code.csv",
@@ -24,6 +24,7 @@
 
         # Accesses
         "security/ir.model.access.csv",
+        "security/ir_rule.xml",
 
         # Views
         "views/product_template.xml",
