+++ addons/l10n_id_efaktur_coretax/controllers/download_efaktur.py
@@ -19,7 +19,7 @@ class EfakturDownloadController(http.Controller):
 
     @http.route('/l10n_id_efaktur_coretax/download_attachments/<models("ir.attachment"):attachments>', type='http', auth='user')
     def download_invoice_attachments(self, attachments):
-        attachments.check_access_rights('read')
+        attachments.check_access('read')
         assert all(attachment.res_id and attachment.res_model == 'l10n_id_efaktur_coretax.document' for attachment in attachments)
         if len(attachments) == 1:
             headers = _get_headers(attachments.name, attachments.mimetype, attachments.raw)
