+++ addons/l10n_id_efaktur_coretax/models/res_partner.py
@@ -1,11 +1,11 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo import fields, models
+from odoo import api, fields, models
+from odoo.addons.l10n_id_efaktur_coretax.models.account_move import TAX_TRANSACTION_CODE
 
 class Partner(models.Model):
     _inherit = "res.partner"
 
-    l10n_id_kode_transaksi = fields.Selection(selection_add=[('10', '10 Other deliveries')], default='04')
     l10n_id_tku = fields.Char(
         default="000000",
         string="TKU",
@@ -20,3 +20,17 @@ class Partner(models.Model):
         ('Other', 'Others')
     ], default='TIN', string="Document Type")
     l10n_id_buyer_document_number = fields.Char(string="Document Number")
+    l10n_id_nik = fields.Char(string='NIK')
+    l10n_id_pkp = fields.Boolean(string="Is PKP", compute='_compute_l10n_id_pkp', store=True, readonly=False, help="Denoting whether the following partner is taxable")
+    l10n_id_kode_transaksi = fields.Selection(
+        selection=TAX_TRANSACTION_CODE,
+        string='Invoice Transaction Code',
+        help="he first 2 digits of tax code",
+        default='04',
+        tracking=True,
+    )
+
+    @api.depends('vat', 'country_code')
+    def _compute_l10n_id_pkp(self):
+        for partner in self:
+            partner.l10n_id_pkp = partner.vat and partner.country_code == 'ID'
