+++ addons/l10n_id_efaktur_coretax/security/ir_rule.xml
@@ -2,7 +2,7 @@
 <odoo noupdate="1">
     <record id="efaktur_document_multi_company" model="ir.rule">
         <field name="name">E-Faktur document multi-company</field>
-        <field name="model_id" ref="model_l10n_id_efaktur_efaktur_range"/>
+        <field name="model_id" ref="l10n_id_efaktur_coretax.model_l10n_id_efaktur_coretax_document"/>
         <field name="domain_force">[('company_id', 'in', company_ids)]</field>
     </record>
 </odoo>
