+++ addons/l10n_id_efaktur_coretax/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_l10n_id_efaktur_coretax
+from . import test_l10n_id_efaktur_download
